// klsd/src/enums/router-name.ts
export const enum RouterName {
  Home = "/",
  Browse = "/browse",
  BrowseDetail = "/browse-detail",
  Search = "/target",
  Target = "/target",
  Compound = "/compound",
  CompoundPrediction = "/compound-prediction", // 新增
  SearchDetail = "/search-detail",
  Drugs = "/drugs",
  DrugsDetail = "/drugs-detail",
  Favorites = "/favorites",
  FavoritesDetail = "/favorites-detail",
  More = "/more",
  MoreDetail = "/more-detail",
  Faq = "/faq",
  DataChart = "/datachart",
  News = "/news",
  ContactUs = "/contactus",
  Personal = "/personal",
  Setting = "/setting",
  PersonalData = "/personal-data",
  SignIn = "/sign-in",
  SignUp = "/sign-up",
  Error = "/404",
  SignOut = "0",
}